moved scanner module from main.rs into different files
