+++ src/scanner.rs
@@ -1,0 +1,4 @@
+pub mod scanner;
+mod token;
+
+pub use scanner::Scanner;
